updated useEffects
