+++ src/components/app/app.tsx
@@ -36,13 +36,7 @@ import { OnlyAuth, OnlyUnAuth } from '../routes/protected-route';
 
 import { fetchIngredients } from '../../services/slices/ingredients-slice';
 import { fetchFeeds } from '../../services/slices/feed-slice';
-import {
-  getUserThunk,
-  selectUser,
-  checkUserAuth,
-  selectAuthCheck,
-  setAuthCheck
-} from '../../services/slices/user-slice';
+import { checkUserAuth, setAuthCheck } from '../../services/slices/user-slice';
 
 function App() {
   const location = useLocation();
@@ -54,12 +48,7 @@ function App() {
 
   const onClose = () => navigate(-1);
 
-  const user = useSelector(selectUser);
-
-  const authChecked = useSelector(selectAuthCheck);
-
   useEffect(() => {
-    dispatch(getUserThunk());
     dispatch(fetchIngredients());
     dispatch(fetchFeeds());
   }, [dispatch]);
@@ -71,7 +60,7 @@ function App() {
         console.log(e);
       })
       .finally(() => dispatch(setAuthCheck(true)));
-  }, [authChecked]);
+  }, []);
 
   return (
     <div className={styles.app}>
