+++ src/services/slices/order-slice.ts
@@ -1,4 +1,4 @@
-import { orderBurgerApi } from '@api';
+import { getOrderByNumberApi, orderBurgerApi } from '@api';
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import { TOrder } from '@utils-types';
 
@@ -6,17 +6,24 @@ interface OrderState {
   success: boolean;
   isLoading: boolean;
   data: TOrder | null;
+  orderByNumber: TOrder | null;
 }
 
 const initialState: OrderState = {
   success: true,
   isLoading: false,
-  data: null
+  data: null,
+  orderByNumber: null
 };
 
 export const orderBurger = createAsyncThunk(
   'order/orderBurger',
   async (ingredients: string[]) => await orderBurgerApi(ingredients)
+);
+
+export const fetchOrderByNumberThunk = createAsyncThunk(
+  'orders/fetchOrderByNumber',
+  async (number: number) => await getOrderByNumberApi(number)
 );
 
 const orderSlice = createSlice({
@@ -45,6 +52,19 @@ const orderSlice = createSlice({
         state.data = action.payload.order;
       })
       .addCase(orderBurger.rejected, (state) => {
+        state.isLoading = false;
+        state.success = false;
+      })
+
+      .addCase(fetchOrderByNumberThunk.pending, (state) => {
+        state.isLoading = true;
+      })
+      .addCase(fetchOrderByNumberThunk.fulfilled, (state, action) => {
+        state.isLoading = false;
+        state.success = true;
+        state.orderByNumber = action.payload.orders[0];
+      })
+      .addCase(fetchOrderByNumberThunk.rejected, (state) => {
         state.isLoading = false;
         state.success = false;
       });
